+++ src/pages/index.page.tsx
@@ -1,4 +1,5 @@
-import Button from '@/components/Button';
+import Button from '@/components/Button/Button';
+import ButtonDashboard from '@/components/Button/ButtonDashboard';
 import Head from 'next/head';
 import Image from 'next/image';
 import Link from 'next/link';
